fix!: markdown not being formatted by prettier
eslint-plugin-format dependency has been replaced with eslint-plugin-prettier.
Because eslint/markdown uses its own parser, the prettier parser did not work
therefore markdown files did not get linted. Currently, eslint-plugin-format does
not have support for non-js native files that don't use the `Program` root.

The consequence of this is that dprint is no longer supported for formatting, but
this was currently only used as an alternative to prettier anyway, and so I don't
imagine anyone was using this functionality anyway. Hopefully this wont be an
actual breaking change for anyone!